update: replace createMachine with interpret
