+++ packages/core/lib/index.ts
@@ -1,7 +1,6 @@
-export { createMachine } from "./createMachine";
 export { MachineConfig } from "./MachineConfig";
 export { createStates, createEvents, createContext } from './utils'
 export { assign } from './assign'
-export { interpret } from './interpret/interpret'
+export * from './interpret/'
 export * from './types'
 export const SSM_VERSION = "0.0.23"
