+++ packages/core/lib/interpret/index.ts
@@ -1,0 +1,2 @@
+export { interpret } from './interpret'
+export type {TReturnState} from './types'
