+++ packages/core/lib/types.ts
@@ -34,9 +34,3 @@ export type TCallback<IContext, IEvents extends IDefaultEvent> = (context: ICont
 export type TAsyncCallback<IContext> = (context: IContext) => Promise<any>
 
 export type TAfterCallback<IContext> = (context: IContext) => number;
-
-export type TCurrentState<U extends TDefaultStates, V> = {
-    value: U[number];
-    history: U[number];
-    context: V
-}
